Move the folders back into main view
Got lots more tests passing
