+++ Tests/MatterControl.AutomationTests/CreateLibraryFolder.cs
@@ -47,7 +47,6 @@ namespace MatterHackers.MatterControl.Tests.Automation
 			{
 				testRunner.CloseSignInAndPrinterSelect();
 
-				testRunner.ClickByName("Library Tab");
 				testRunner.NavigateToFolder("Local Library Row Item Collection");
 				testRunner.ClickByName("Create Folder From Library Button");
